+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/validation_annotations.php
@@ -9,6 +9,7 @@ $container->loadFromExtension('framework', [
     'validation' => [
         'enabled' => true,
         'enable_annotations' => true,
+        'email_validation_mode' => 'html5',
     ],
 ]);