+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/validation_auto_mapping.php
@@ -7,6 +7,7 @@ $container->loadFromExtension('framework', [
     'php_errors' => ['log' => true],
     'property_info' => ['enabled' => true],
     'validation' => [
+        'email_validation_mode' => 'html5',
         'auto_mapping' => [
             'App\\' => ['foo', 'bar'],
             'Symfony\\' => ['a', 'b'],